+++ src/Table/Actions/DeleteAction.php
@@ -1,0 +1,22 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Milenmk\LaravelSimpleDatatables\Table\Actions;
+
+use Closure;
+
+class DeleteAction extends BaseAction
+{
+    public string|Closure|bool|null $icon = 'heroicon-o-trash';
+
+    /**
+     * Set the label for the column. This can be a string, array, or callable.
+     */
+    public function label(string|array|null $value): self
+    {
+        $this->label = $value ?? __('Delete');
+
+        return $this;
+    }
+}